Clean up L1Block.sol

+++ contracts/src/upgrades/L1Block.sol
@@ -162,9 +162,9 @@ contract L1Block is ISemver, IGasToken {
         uint128 periodMinted;
         uint128 maxSupply;
         uint128 initialTargetPerPeriod;
-        uint128 targetNumBlocksInHalving;
-        uint128 expectedTotalMinted;
-        int128 pacingDelta;
+        uint256 targetNumBlocksInHalving;
+        uint256 expectedTotalMinted;
+        int256 pacingDelta;
     }
     
     function fctDetails() external view returns (FctDetails memory) {
@@ -181,22 +181,22 @@ contract L1Block is ISemver, IGasToken {
         });
     }
 
-    function targetNumBlocksInHalving() public pure returns (uint128) {
+    function targetNumBlocksInHalving() public pure returns (uint256) {
         return 2_628_000;
     }
     
-    /// @notice Returns the expected total minted at current block based on linear schedule
+    /// @notice Returns the approximate expected total minted at current block based on linear schedule
     /// @return Expected total minted amount in wei
-    function targetTotalMinted() public view returns (uint128) {
-        uint128 supplyTargetFirstHalving = uint128(fctMaxSupply) / 2;
-        return (supplyTargetFirstHalving * uint128(block.number)) / targetNumBlocksInHalving();
+    function targetTotalMinted() public view returns (uint256) {
+        uint256 supplyTargetFirstHalving = fctMaxSupply / 2;
+        return (supplyTargetFirstHalving * block.number) / targetNumBlocksInHalving();
     }
     
-    /// @notice Returns pacing delta: positive if ahead of schedule, negative if behind
+    /// @notice Returns approximate pacing delta: positive if ahead of schedule, negative if behind
     /// @return Pacing delta as a percentage (e.g., 0.05e18 = 5% ahead)
-    function pacingDelta() public view returns (int128) {
-        uint256 ratio = (uint256(fctTotalMinted) * 1e18) / targetTotalMinted();
-        int256 delta  = int256(ratio) - int256(1e18);
-        return int128(delta); 
+    function pacingDelta() public view returns (int256) {
+        uint256 ratio = (fctTotalMinted * 1e18) / targetTotalMinted();
+        int256 delta = int256(ratio) - 1e18;
+        return delta; 
     }
 }
